sithafal/Task 2/requirements.txt
#requirements.txt

requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.3
transformers==4.33.2
numpy==1.24.3
